代理反水计算

+++ src/main/java/com/yingliguoji/fee/dao/GameRecordMapper.java
@@ -1,10 +1,11 @@
 package com.yingliguoji.fee.dao;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public interface GameRecordMapper {
 
-    Integer getPlayerTotal(Map<String,Object> data);
+    BigDecimal getPlayerTotal(Map<String,Object> data);
 
 }
